chore: relocate delete category modal
